Implement a wininet backend for the new netcode.
This will allow us to do the transition in 2 steps, by first using the new netcode, and then figuring out how to build libcurl/openssl in our environment.

This was surprisingly fast. As soon as it compiled, it worked. Literally no debug needed. And no need to change the interfaces either (well, I'll need a better way to call curl_global_init on the curl side while doing nothing on the wininet side. Either some static function on HttpHandle, or initialize on the 1st caller. The 2nd one isn't really viable because curl_global_init must be called before creating any thread IIRC).

+++ netcode/file.cpp
@@ -17,7 +17,7 @@ FileStatus File::getStatus() const
     return this->status;
 }
 
-unsigned int File::getThreadsLeft() const
+size_t File::getThreadsLeft() const
 {
     return this->threadsLeft;
 }
